Fix test failures from users section unification
