+++ features/admin-users-v1/components/wizard/steps/add-user-groups.tsx
@@ -159,7 +159,7 @@ export const AddUserGroups: FunctionComponent<AddConsumerUserGroupPropsInterface
                                     listItemId={ group.id }
                                     listItemIndex={ index }
                                     listItemTypeLabel={ null }
-                                    isItemChecked={ checkedUnassignedListItems.includes(group) }
+                                    isItemChecked={ checkedUnassignedListItems?.includes(group) }
                                     showSecondaryActions={ false }
                                     handleOpenPermissionModal={ () => handleSetGroupId(group.id) }
                                     data-testid="user-mgt-add-user-wizard-modal-unselected-groups"
